Better documentation  for CETA + structure change
The documentation for all CETA functions has been updated to include a description of their outputs.
The "not quite CETA" function for trajectory sections have been put in a separate file. The documentation is there too but is not exported (I used @noRd).
Small changes have been made to the code to make it clearer:
-systematize the use of "<-" instead of "=" for attribution.
-harmonize variable names in camelCase throughout the extension (mostly?)
-make more understandable some parameter names.